ChatBot/requirements.txt
huggingface_hub==0.25.2
polygon-api-client
unsloth
gradio